fix: harden findings from the adversarial refactor review
Fixes from a multi-agent review of the whole branch (4 lenses, each
finding adversarially verified against the original firmware):

- CRITICAL stale-now underflow: appRun sampled millis() BEFORE
  modbusServerTick, so handler-stamped timestamps (LED on-time, latch
  DELAY start) were newer than 'now' and the uint32 subtraction
  underflowed — the LED was forced off in the same tick it turned on
  and the configured unlock delay was skipped. 'now' is now sampled
  after the Modbus tick.
- Latch pulse hardware guard (TIM7 one-shot ISR): the 500ms solenoid
  clamp no longer depends on loop cadence — a long Modbus response
  flush or noise-stalled poll() mid-PULSE can no longer stretch the
  pulse past the limit.
- finishRequest syncs the enable coil only while the LED is actually
  on, so coil 1001 can no longer read 1 with the ring dark (which also
  made a subsequent write-1 a shadow-suppressed no-op).
- settingsInit retries the boot read 3x and, on persistent failure,
  runs on RAM defaults WITHOUT formatting — a transient I2C glitch can
  no longer wipe a valid stored configuration.
- Torn-save hardening: routine saves write only payload+CRC (magic is
  never rewritten), first-time format commits the header last, and the
  legacy MCU-flash copy is tombstoned after migration — a power-loss
  tear can corrupt at worst the CRC (identifier-salvage path), never
  resurrect stale pre-migration config.
- Slave ID validated (1-247) at persist time and at boot; the
  modbusServerInit parameter widens to uint16_t. Previously reg 4=300 +
  coil 503 would boot the device answering at ID 44 (300 mod 256) —
  a wrong-cabinet unlock risk on a multi-drop bus.
- IWDG timeout 1s -> 4s: a single poll() can legitimately stall for
  hundreds of ms under bus noise; the latch clamp no longer relies on
  the watchdog.

+++ src/svc/modbus_server.h
@@ -25,9 +25,9 @@ enum MbWatchKind : uint8_t
 typedef void (*MbWatchHandler)(uint16_t addr, uint16_t value);
 
 /*  @brief Start the RTU server on the RS485 port.
- *  @param slaveId Modbus slave ID
+ *  @param slaveId Modbus slave ID (validated range 1-247)
  *  @param baud    UART baud rate (must match rs485PortBegin) */
-void modbusServerInit(uint8_t slaveId, uint32_t baud);
+void modbusServerInit(uint16_t slaveId, uint32_t baud);
 
 /*  @brief Poll the bus; on activity, dispatch the watch table. */
 void modbusServerTick();
